added exception handler

+++ src/main/java/com/example/bookmarkmanager/services/BookmarkService.java
@@ -1,5 +1,7 @@
 package com.example.bookmarkmanager.services;
 
+import com.example.bookmarkmanager.exception.Exception;
+import com.example.bookmarkmanager.exception.ResourceNotFoundException;
 import com.example.bookmarkmanager.model.Bookmark;
 import com.example.bookmarkmanager.model.Folder;
 import com.example.bookmarkmanager.repositories.BookmarkRepository;
@@ -10,7 +12,6 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 @Service
 public class BookmarkService {
@@ -35,9 +36,14 @@ public class BookmarkService {
     }
 
     public void deleteBookmark(Long bookmarkId) {
+
+       if (bookmarkId == null) {
+           throw new ResourceNotFoundException(
+                   "Please supply bookmark id");
+       }
         boolean exists = bookmarkRepository.existsById(bookmarkId);
         if (!exists) {
-            throw new IllegalStateException(
+            throw new ResourceNotFoundException(
                     "Bookmark with id = " + bookmarkId + " does not exists");
         }
         bookmarkRepository.deleteById(bookmarkId);
@@ -52,28 +58,27 @@ public class BookmarkService {
         //            "Bookmark with name = " + bookmark.getName() + " is exists");
        // }
 
-       if (bookmark.getFolder() == null) {
-            bookmarkRepository.save(bookmark);
-        } else {
-            if (bookmark.getFolder().getId() == null) {
-                throw new IllegalStateException(
-                        "Invalid JSON format"
-                );
+            if (bookmark.getFolder() == null) {
+                bookmarkRepository.save(bookmark);
+            } else {
+                if (bookmark.getFolder().getId() == null) {
+                    throw new Exception("Please specity folder Id");
+                }
+
+                boolean exists = folderRepository.existsById(bookmark.getFolder().getId());
+                if (!exists) {
+                    throw new ResourceNotFoundException(
+                            "Folder id = " + bookmark.getFolder().getId() + " does not exists");
+                }
+                Folder folder = folderRepository.findById(bookmark.getFolder().getId()).orElse(null);
+                if (null == folder) {
+                    folder = new Folder();
+                }
+                folder.setId(bookmark.getFolder().getId());
+                bookmark.setFolder(folder);
+                bookmarkRepository.save(bookmark);
             }
 
-            boolean exists = folderRepository.existsById(bookmark.getFolder().getId());
-            if (!exists) {
-                throw new IllegalStateException(
-                        "Folder id = " + bookmark.getFolder().getId() + " does not exists");
-            }
-            Folder folder = folderRepository.findById(bookmark.getFolder().getId()).orElse(null);
-            if (null == folder) {
-                folder = new Folder();
-            }
-            folder.setId(bookmark.getFolder().getId());
-            bookmark.setFolder(folder);
-            bookmarkRepository.save(bookmark);
-        }
     }
 
     @Transactional
@@ -82,18 +87,18 @@ public class BookmarkService {
                                String url,
                                Long folderId) {
         Bookmark bookmark = bookmarkRepository.findById(bookmarkId).
-                orElseThrow(() -> new IllegalStateException("Bookmark with id = " + bookmarkId + "does not exists"));
+                orElseThrow(() -> new ResourceNotFoundException("Bookmark with id = " + bookmarkId + " does not exists"));
 
-       // Optional<Bookmark> bookmark1 = bookmarkRepository.findBookmarkByName(name);
+        // Optional<Bookmark> bookmark1 = bookmarkRepository.findBookmarkByName(name);
         //if (bookmark1.isPresent() &&
-         //       !Objects.equals(bookmark1.get().getId(), bookmarkId))
-          //     {
-           // throw new IllegalStateException("Bookmark name =" + name + " belongs to bookmark id = " + bookmark1.get().getId());
-       // }
+        //       !Objects.equals(bookmark1.get().getId(), bookmarkId))
+        //     {
+        // throw new IllegalStateException("Bookmark name =" + name + " belongs to bookmark id = " + bookmark1.get().getId());
+        // }
 
-       if (folderId != null ) {
+        if (folderId != null) {
             Folder folder = folderRepository.findById(folderId).
-                    orElseThrow(() -> new IllegalStateException("Folder id = " + folderId + "does not exist"));
+                    orElseThrow(() -> new ResourceNotFoundException("Folder id = " + folderId + " does not exist"));
             bookmark.setFolder(folder);
         }
 
@@ -110,5 +115,4 @@ public class BookmarkService {
         }
 
     }
-
 }
